+++ Assets/Animations/Tutorial.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 35d22f6da443fc24185ca76568af94d0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
