+++ Assets/Animations/Tutorial/CloseTutorial.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 898aa649872f5bb4789ea31ff0102097
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
