+++ Assets/Animations/Tutorial/Tutorial.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: beeb8a2da8f02c24db58690e70a6ef4d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
