--- Assets/Scripts/Quiz/ButtonsController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 39d96594584b49d43bc941c6ae9f5f6a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
