+++ Assets/Scripts/TutorialController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 34f4dd9a8ff4ec8408ab40e23ebd4fae
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
