--- Assets/Tilemap/Tiles/RuleTile/New Animation.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b5be4aeaa56012842b2ef19e8671c797
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
